buck macros: set `FOLLY_XLOG_STRIP_PREFIXES` in modular builds for xlog
Summary:
In modular builds, we compile a rules headers into a module using a Buck
`cxx_genrule`, which sandboxes it's sources under a directory under buck-out.
Since the location of this output directory doesn't match the heuristics used
by xlog for it's default buck-out stripping, these full buck-out paths show up
during logging, instead of human-friendly fbcode-relative paths.

This diff uses xlog's existing support for `FOLLY_XLOG_STRIP_PREFIXES` to fix
this, and has each module compilation pass in it's rule/file-specific sandbox
location, so that xlog with strip this off.

Currently, there's some issues with how `FOLLY_XLOG_STRIP_PREFIXES` is handled
in modular builds.  The xlog.h, header which implements `FOLLY_XLOG_STRIP_PREFIXES`
support, is compiled into it's own module, and so value of
`FOLLY_XLOG_STRIP_PREFIXES` at the top-level refers to prefix specific to this
file's owning rule, when what we really care about is the value of this  macros
at the logging callsite, in downstream dependents.  To work around this, this
diff avoids the `#ifdef FOLLY_XLOG_STRIP_PREFIXES` at the top-level of xlog.h
when modules are enabled.

Another issue is that when setting `FOLLY_XLOG_STRIP_PREFIXES` only for module
compilations, the modular build of xlog.h originally has this macro set, but it's
imported into contexts that don't, so the resulting macro definition of
`XLOG_FILENAME` attempts to again read from `FOLLY_XLOG_STRIP_PREFIXES` and
fails.  There might be a way to make this work w/ and w/o `FOLLY_XLOG_STRIP_PREFIXES`
set, but I'm not aware of this.

To work around this second issue, also defined `FOLLY_XLOG_STRIP_PREFIXES` in
non-modular contexts.  This value is only really useful for modular builds, as
in non-modular builds, we'd still need to pass in header symlink dirs of all
transitive deps.

Reviewed By: pixelb

Differential Revision: D13079213

fbshipit-source-id: ba2ae39f694def61446d430fc7e0d2e6592b644e