image_package can emit btrfs images (for loopback mounts)
Summary:
This is the "prototype-grade" loopback output implementation. I have a more production-ready version in the pipeline, but I didn't want to block people from playing with loopback images while I write tests for the new set of modules.

To help reviewers focus on the "good parts" of this code, here is what will change in the production code:
 - `btrfs_loopback.py` will be completely replaced by tested, more robustly structured code — that's why it has no dedicated tests right now
 - `subvol_utils.py` will be updated to use the new APIs, and will get some more tests, but will not change much in substance.
 - the `package_image.py` implementation is essentially final

Reviewed By: marcinpe

Differential Revision: D13213360

fbshipit-source-id: 1b7b3193beccbda303f7d954a5b06219f75438c3