Remap no_aop guice to regular guice
Summary:
Conflicts between no_aop and regular guice artifacts were causing
runtime errors on Sandcastle for D13202393 (e.g. P60379226). This
diff adds the option to remap artifacts with a classifier and uses
it to remap the no_aop guice artifact (which is only included in
third-party-java indirectly) to the regular artifacts.

Differential Revision: D13244694

fbshipit-source-id: 90e06b61b32342fde2ae271cf36b00fc13e1b74a