buck macros: restore separate extended build mode names in test console
Summary:
We're not setting the sanitizer label for `dev` which is making tests naming
collide in test console.  On top of this, test console emits a repro using the
build mode named in the label we pass in, which means we need to pass the build
mode as written on the command line.

Reviewed By: vrishal, markisaa

Differential Revision: D13289305

fbshipit-source-id: 07ce8d1bdf4f661ace77a74c94d41d828c90e24b