buck macros: workaround overriden module home path version suffix bug
Summary:
When a library building a module has a versioned (transitive) dep which would
otherwise cause a version suffix to be used and if this library is also specified
on the command-line explicitly, this creates a situation where the library will
not use a version suffix, but the `cxx_genrule` we create to build it's module
will.  Since we embed the overridden module home in the `cxx_genrule`, we use
it's version suffix, which is almost always the same as the top-level library's,
when forming the overridden path.  This means, in the scenario described above,
we embed the wrong version suffix into the module, which causes errors when the
module tries to run through verification checks in clang (resulting in it being
unable to locate input files properly).

This diff works around this issue by using the `-fmodules-embed-all-files` to
prevent this module input file verification, as it shouldn't be necessary anyway.

Reviewed By: igorsugak

Differential Revision: D13317678

fbshipit-source-id: 7cfb311d111c28cdb561050b901a97ebe17b2f38