thrift_library() py3 generated .thrift sources fix
Summary:
When we recieve a thrift_srcs, it was not already normalized like it is
when we are passed a thrift_src. That is the generated part of the
thrift_src being removed.

We re-create the thrift_srcs dict with the .thrift file normalized,
since we do lookups on this dict using a normalized name.

Reviewed By: JonCoens

Differential Revision: D13411461

fbshipit-source-id: 35a4b289c2d6b6f17a63c90e0a6ba0b7c42a40de