Remove some deduping logic by just not doing the duplicate work
Summary: Before we would generate the gen_test_modules library for every single versioned binary when using helpers, then we deduped it. Since we don't return this information in skylark, pass in a 'generate_test_modules' flag that only gets set on the first library, and remove the dedupe logic. This will probably be cleaned up a bit more in a future diff.

Reviewed By: andrewjcg

Differential Revision: D13442531

fbshipit-source-id: 003ef62f0ab6983c231e1960a64934215397f70d