Add python_version method to replace PythonVersion
Summary: This creates an analog to PythonVersion as found in python.py. The major difference here is that we don't quite work like LooseVersion (and that shows in the subsequent migration of PythonVersionConstraint where we just compare major/minor/patch + flavor). I'll verify but I don't /think/ we rely on any odd behavior where we expect LooseVersion("2") === LooseVersion("2.0")

Reviewed By: andrewjcg

Differential Revision: D13433401

fbshipit-source-id: ec8922ae3dc8d87391bca794009d34fe5f158de7