fbcode macros: add virtual cells abstraction to handle tp2 dep translation
Summary:
In fbcode, we support a `third-party//` "virtual" cell which maps to a Buck
C/C++ platform-specific third-party2 root (e.g. `third-party-buck/platform007/`).
Currently, most of this logic is pretty specific to fbcode's toolchain platforms
and third-party2 setup.

This diff adds a new "virtual cell" abstraction so that this translation logic
which is fbcode-specific can be hidden behind the new `CxxPlatformInfo` type,
and updates the macro lib to use this when formatting parameters. This will
allow other `CxxPlatformInfo` types (e.g. for "macosx-x86_64") to implement
their own translation strategy (e.g. redirect to `xplat//third-party`) without
changes to the macro lib.

Reviewed By: pixelb

Differential Revision: D13396659

fbshipit-source-id: dd3879c2a7658189245f3774b6fefc8e751e139d