buckify_tp2: decouple missing project stubs from platforms
Summary:
Previously, we fill in "missing" stubs for tp2 projects which didn't exist on
a given platform.  This diff moves these stubs to a platform-agnostic location
so that any platform (e.g. platform007, macos-x86_64) can redirect to them
instead of generating and using platform-specific stubs.

This does mean we're doing more work in the script, however, this is hidden by
moving all of it into a parallel task (via multiprocessing), such that this
doesn't impact the overall run time of buckify_tp2.

Reviewed By: pixelb

Differential Revision: D13421913

fbshipit-source-id: a066e5fffe4d9d0d63641ffcc9de672a79d2d4c0