fbpkg_builder: A Buck rule to help build fbpkgs without Configerator
Summary:
== Intro

Most of this diff summary attempts to answer questions like "why this?" and "what are your plans?". To get an idea of what this does, take a look at `fbcode/fs_image/buck_macros/facebook/tests/TARGETS`. To build one of the fbpkgs from that `TARGETS` file, you would run:

```
$ buck run //fbcode/fs_image/buck_macros/facebook/tests:fbpkg_builder_test_package_with_symlink
... stderr spam ...
fbpkg_builder_test_package_with_symlink:daaff4cca05f4878b0689699421ad8bb
```
For the interface & implementation specifics, please read the detailed docblock.

== What will this be used for?

Right now, I have to create ~80 fbpkgs for OS images in fbcode, which would translate to 80 configerator diffs + 80 fbcode diffs. The cross-repo split is shaping up to be maintenance nightmare, since as the rollout of images proceeds, we will need to vary what goes in the packages and how many packages there are.

What I have to do now:
 - fbcode business logic: D13274501
 - configerator boilerplate, 34 significant lines: D13274984

With this new facility, I just need 80 fbcode diffs, adding a short section to each TARGETS file, equivalent to:

```
fbpkg_builder(
  name = "tupperware.image.webfoundation.facebook.com",
  path_actions = {
    "image.diff": ("copy", ":facebook.com.sendstream"),
    "image.btrfs": ("copy", ":facebook.com.btrfs"),
  },
)

fbpkg_builder(
  name = "tupperware.image.webfoundation.facebook.com.twshared",
  path_actions = {
    "image.diff": ("copy", ":facebook.com.twshared.sendstream"),
    "image.btrfs": ("copy", ":facebook.com.twshared.btrfs"),
  },
)
```

In actuality, I'll wrap this with a TW-specific macro to cut the fbpkg-creation boilerplate to **1 line per fbpkg**.

Next half, we expect customers to create many more new OS images, so this is a very meaningful usability win.

== What about use-cases other than OS images?

This is in `fbcode/fs_image` because I don't plan to advertise this in H1. For now, it's a non-priority to drive adoption for this macro. If my experience with this goes well, I can make a post in H2 and see if other people want it.

But, during the design and implementation, keeping this generally useful was a core guiding principle — so I hope it didn't turn out too badly, and can help others.

Here is the relevant "why?" from the "Rationale" section:

Q:  Why should I use this instead of `fbpkg build` with traditional Configerator BuildConfigs?
A:  You don't have to use this, but some benefits include:
 - Developer efficiency: Adding a new file to your service? You only need one diff in one repo.
 - Determinism: Everything about your fbpkg version is determined by an fbsource hash (caveat: `fbpkg meta` is globally mutable),

== Why am I listed as a reviewer?

 - markisaa, Ben0mega because you work on contbuild and have contributed to the fbpkg integration
 - mkacik because I plan to copy-pasta your "fbpkg bundles for contbuild" work for this feature
 - andrewjcg because you wrote `fbpkg.cinc` — I was wondering if I was missing anything by explicitly setting build-info-related configs the way that `fbpkg.cinc` does? (D2910038)
 - jhunt and bsbates as an FYI regarding their work on hermetic configs & fbpkgs.
 - cjhopman and philipjameson since they had graciously helped me iterate on making this new macro a sane member of the Buck ecosystem.

Reviewed By: zeroxoneb

Differential Revision: D13479607

fbshipit-source-id: 4651ed17c6f3061172d65d389bead6f701d2af50