Add support for symlinks to buck images
Summary:
Add support for symlinks to buck images.

This adds the necessary plumbing to implement symlinks to directories and files that are contained within the image. This does not support dangling symlinks, the source file/directory must exist inside the image for the symlink to be created.

The description of how to use this is included in the doc block of  `image_feature.bzl`.

Reviewed By: snarkmaster

Differential Revision: D13576046

fbshipit-source-id: e3bc1ce85e56052f4e51c486392429811b84c581