Add `remove_paths` functionality
Summary:
We need to be able to clean out RPM-related unnecessary files from our resulting images. Up next: an `image.feature` that removes `/var/cache/yum`.

Fear not: tests comprise most of this diff, the implementation is pretty simple.

The testing burden is higher than for regular "additive" features since this is a phase, and so we should cover interactions with prior phases as well as later non-phase items.

For the description of the new functionality, I will just copy-paste the inline comment:

> An iterable of paths to files or directories to (recursively)
> remove from the layer.  These are allowed to remove paths
> inherited from the parent layer, or those installed by RPMs even
> in this layer.  However, removing other items explicitly added by
> the current layer is currently not supported since that seems like
> a design smell -- you should probably refactor the constituent
> image.feature`s not to conflict with each other.
>
> By default, it is an error if the specified paths are missing from
> the image.  This form is also supported:
>      `[('/path/to/remove', 'assert_exists|if_exists')],`
> which allows you to explicitly ignore missing paths.

Reviewed By: zeroxoneb

Differential Revision: D13794153

fbshipit-source-id: 81facc046c5652576d1ba9377cf73d7a965594ac