Make Subvol.path more paranoid about out-of-subvolume symlinks
Summary:
The main goal of this change is to draw more attention to the following potential bug category:
 - Some of the image build operations use un-chrooted utilities to write to the image.
 - These utilities will commonly follow existing symlinks inside the image when writing (e.g. `tar` writes through existing symlinks and there is no option to block that).
 - A symlink inside the image, as interpreted by the un-chrooted utility, can end up pointing outside the image. Specific examples include symlinks to absolute paths, and broken symlinks. Writing through such a symlink could make the image build step make errant writes outside of the image.

The explicit concern here is producing incorrect images. Image configurations are trusted code, so we are not trying to ward off malice, but only accidents.

The long-term fix would be to either chroot all utilities into a read-only bind-mount of the host filesystem, or better yet to run them from a read-only, deterministic "appliance" image with host-independent versions of the utilities.

However, for the time being, let's make the path handling more robust to such failures.

I should note that I'm not aware of any configs that would cause an `image.layer` to be constructed incorrectly, and to errantly write outside the container. So this patch is part of a "defense in depth" approach to preventing bugs. The specific thing that prevents all known shenanigans is that the image compiler explicitly validates its filesystem actions before executing them.

Reviewed By: ryantimwilson

Differential Revision: D13789277

fbshipit-source-id: 865bb75cdeb1f5f2d5d4367ad1bbcd0a5a51dd59