Make `image.layer` mountable inside `image.layer`
Summary:
Mechanistically, this changes the output type of `image.layer` from `blabla.json` to `blabla.layer/{layer.json,mountconfig.json}` where `layer.json` is the prior output, and `mountconfig.json` contains the information required to mount this target.

To understand what mounting a target means, read D14019576 that adds `mounts` to `image.feature`. This here is the simplest useful example of a mountable target type. This diff also hints how images can easily be assembled out of sub-images.

NB: The lack of a `runtime_source` implementation means that this mount is currently available in a Buck repo, but not in packaged images running in production. This could be improved.

Reviewed By: zeroxoneb

Differential Revision: D14019575

fbshipit-source-id: 9cd2844928ba7db4bcef3bfa49bdca3601e16e57