Add support for host mounts, mounts work with files OR directories
Summary:
In general, using host mounts in containers is a undesirable implementation crutch, since it makes containers non-hermetic, and creates potentially dangerous coupling between the host and the container.

In some specific settings, this is a useful crutch to have. The reason I am adding support for host mounts right now is that I need host mounts in order to make it possible for developers execute Buck's "in-place binaries" after copying their very non-relocatable binaries into the container.

A quick re-cap of how "in-place binaries" operate: when a `python_binary` is built in `mode/dev`, its output is just a tiny bootstrap script that runs the Python interpreter against a "link-tree". The link-tree is a directory that gets output alongside with the `python_binary`, and contains symlinks to the actual `.py` files in the source repos, arranged according to their `base_module` declarations. The upshot is that in order to execute a `mode/dev` binary, we MUST have the source tree available in the container, at its current real path. We also need the host's "runtime" libraries that were used to build the binary. This is where container host mounts come in.

The above reasoning applies to C++ as well, except that in this case, the binary is dynamically linked against `.so` files produced by `cpp_libraries` built in the repo.

In order to properly support host mounts, this also changes `mounts` to use read-only rslave rbind mounts — the rationale is in a code comment. This should not affect the previous `mounts` use-cases.

The fact that I am **also** adding support for file mounts in this diff is a bit incidental, but I basically wanted to be able to mount host files as well as directories. If the reviewers protest, I can spend some time to split this into two changes.

Reviewed By: zeroxoneb

Differential Revision: D14193372

fbshipit-source-id: e417ecf281aa5f3fe1a34072a4a3160766f49966