`nspawn_in_subvol` defaults to running as `nobody`; add `/logs` directory.
Summary:
Since we want lots of people to write tests using this runtime, there will be a lot fewer broken windows in our future if containers default to running as an unprivileged user. Plus, it encourages a better security posture for production.

Note that this changes the behavior of `-container` targets to be unprivileged by default, too.

The `/logs` directory is just something that users of the runtime are likely to assume, so I'm baking it in, however inelegantly.

Differential Revision: D14278244

fbshipit-source-id: 034d934a85066611ff6d1c5411dccdd1e44d3311