Improve environment variable forwarding through `sudo`
Summary:
This change was prompted by the observation that `sudo`'s internal forwarding of environment variables is entirely at the mercy of the local configuration. There are a few places where environment variables are important for the code working sanely, loosely:
 - `UNIFIED_CGROUP_HIERARCHY` for CI containers
 - `PYTHONDONTWRITEBYTECODE` to avoid polluting `buck-out` with `root`-owner `.pyc` files.

In view of this, I just put those variables behind the `env` wrapper, which will either fail loudly, or work.

Reviewed By: marcinpe

Differential Revision: D14278245

fbshipit-source-id: 07334bc3195ab63917a0d74b6371b5dfe69be343