Drop `Details.*` from header suffixes
Summary:
We use a "related header suffixes" list to find headers which correspond to
a source that should be owned by the same rule when using `AutoHeaders.SOURCES`
(the default).

This diff begins cleaning this list up a bit, by removing the `Details*` suffixes, as
these are mostly unintuitive for users and slow down Buck parsing and autodeps since
it requires additional globbing.

Updates `tools/build/bin` to `fbbuild:1829`.

Reviewed By: philipjameson, luciang

Differential Revision: D14934763

fbshipit-source-id: 2989065fb80bfbc3b04258f6176034741cfc3eb6