Drop `If.*` from header suffixes
Summary:
We use a "related header suffixes" list to find headers which correspond to
a source that should be owned by the same rule when using `AutoHeaders.SOURCES`
(the default).

This diff begins cleaning this list up a bit, by removing the `If*` suffixes, as
these are mostly unintuitive for users and slow down Buck parsing and autodeps since
it requires additional globbing.

Updates `tools/build/bin` to `fbbuild:1831`.

Reviewed By: philipjameson, luciang

Differential Revision: D14936480

fbshipit-source-id: a2f2bc45f192086445bb06f908fca5af94de12e7