fix dependancy graph for is_repo_root
Summary:
A MountItem with is_repo_root creates all parent dirs for the mount point: see mkdir_opts.append('-p') in build_resolves_targets().
This cheats gen_dependency_order_items() because dirname(mountpoint) may not be provided by any item and hence the MountItem not to be dependent on the ParentLayerItem.
The result is a flakiness of building fb_build_appliance_testing: if ParentLayerItem comes first, everthing is alright, otherwise:
```
[CONTEXT]     assert yield_idx == 0, f'{item}: the parent layer must be 1st'
[CONTEXT] AssertionError: ParentLayerItem(from_target='fake', path='/mnt/btrfs/trunk-hg-fbcode-fbsource-51208074-1556313572/fbcode/buck-image-out/volume/targets/fb_build_appliance_testing:JDxaU8s.Cfw.VU_7/volume'): the parent layer must be 1st
```

Reviewed By: zeroxoneb

Differential Revision: D15132273

fbshipit-source-id: 8e0d1f2b68d96f6727e6be213c80bb8db1fb96bc