Remove /dev/fuse as an 'fb_infra_support' host mount.
Summary:
The main reason for this is that it breaks `yum-from-repo-snapshot`,
the reason being that when we start yum-from-repo-snapshot we explictly bind
mount an empty `/dev` in place to protect it from yum's evil clutches.  Since
we have an empty `/dev` we cannot bind mount `/dev/fuse` in place because the
target file does not exist.  We really probably don't need to setup `/dev/fuse`
as a host mount anyway because 1) the agent will create `/dev/fuse` as part of
the container setup since we require it for `xar` files and 2) we explicitly
create this bind mount in `nspawn-in-subvol` because of the need for `xar`.
So I think until we have a better way of handling `/dev` in
`yum-from-repo-snapshot` or we have a more explicit and pressing need to setup
`/dev/fuse` as part of the `fb_infra_support` should remove this.

Also, this fixes things for: D15247237

Reviewed By: snarkmaster

Differential Revision: D15371904

fbshipit-source-id: b0dc05d64830ab4cf8d2c18715d6e379de29dc18