`copy_deps` no longer supports the rsync convention
Summary:
Right now, `copy_deps` supports the "rsync convention". This feels convenient, but at the end of the day, in Buck, you cannot know the output filename of the target you are referencing. It can be changed under the hood. For example, it's not impossible that all `python_binary` targets would start outputting `.xar` extensions. That specific example is unlikely due to other backwards-compatibility issues, but it illustrates the point that the "rsync convention" is a bad API.

I see these three options:
 - Leave it as-is — this is sure to cause users grievous surprises in the future.
 - Change it to use the target name instead of the target output filename — saves keystrokes, wysiwig, but still not what users would expect.
 - Require explicit names always.

I think we basically _have_ to go with the last option. At any rate, by eliminating this shorthand now, this creates the opportunity of allowing trailing slashes in the future with whatever semantics is deemed to be good.

Reviewed By: zeroxoneb

Differential Revision: D15956248

fbshipit-source-id: 84075ce248ee0b83d20a649aab9e11657d3de826