Prevent bugs with Buck's `$TMP` being used under `sudo`
Summary:
The new comment in the code is all you really need to know. There were real bugs happening due to this, and this should prevent the whole class of them.

Now, this fix is a little-bit heavy-handed in that it might have been "better" to ask whether `$TMP` is in fact pointing into `buck-out`, and only sanitize it in that case. However, the logic for this is kind of messy, and I don't yet know of a case where we would want to pass **other** `$TMP` settings down through `sudo`.

For those who are not faint of heart, here is a mini-repro of the `systemd-nspawn` bug that I was triggering in the building of the `validate-build-appliance` layer upon introducing D16002088:

```
$ pushd $(mktemp -d)
$ mkdir ro
# NB: This can be any btrfs subvol that looks like an OS tree to nspawn:
$ some_os=/home/lesha/fbcode2/buck-image-out/volume/targets/fb_build_appliance:JFvBrQw.KDr8.h0I8/volume
$ TMP=$PWD/ro sudo systemd-nspawn -x --directory "$some_os" --bind-ro $PWD/ro
...
Press ^] three times within 1s to kill container.
Failed to create inaccessible file node '/tmp/tmp.egvrIaMtTL/ro/.#inaccessible04f9cda5452fcda2': Read-only file system
...
```

In other words, `nspawn` was using a `$TMP` path that came from the host **inside** the container environment, where it was pointing at a read-only directory. Oops.

This specific error message was deleted in poettering's OCI mega-diff:

https://github.com/systemd/systemd/commit/d0b6a10c005ab1fff44d032be995e3f2bcfba225

That said, I don't really know if Buck's weird TMP settings would cause other problems after systemd v242.

Reviewed By: zeroxoneb

Differential Revision: D16002853

fbshipit-source-id: 3a5983615eff26a0657571de8d6e49545cd4b573