`nspawn_in_subvol` decides when to bind-mount the repo based on the image metadata
Summary:
Prior to this diff, we had a handful of ways to allow execution of Buck-built in-place binaries (`mode/dev`) from within images:
  - `image.python_unittest` had its own check,
  - `-container` targets for `image.layer`s had their own check,
  - the testing build appliance had yet another mechanism.

In a later diff, I caught myself needing to add a FOURTH such check.

Then it occurred to me that whether we need to bind-mount the repo at runtime is a simple property of the **image** that is being nspawned into. And that it's actually not adequate to ask whether the ambient Buck run would produce in-place binaries, because — I quote from a comment on this diff —

> In the case of build appliance images, it is possible for a mode/dev (in-place) build to use **either** a mode/dev, or a mode/opt (standalone) build appliance. The only way to know to know if the appliance needs a repo mount is to have a marker in the image.

So this diff adds a flag `--artifacts-require-repo` to the compiler (docs in the code), which deposits a boolean into `meta/private/opts/artifacts_require_repo`.

Then, `nspawn_in_subvol` reads this from the subvolume that's being executed.

The rest of the diff just eliminates the prior special-cases, and updates the unit tests.

NB: I remove the actual implementation of `is_repo_root` in a later diff, just to keep things reviewable. Considered from the point of view of significant lines / amount of logic before & after, this change is a significant simplification.

Reviewed By: zeroxoneb

Differential Revision: D16002088

fbshipit-source-id: fa0806407cddee4e3a95f1898fd5dede92ca53f2