Database access module for storing repo objects in MySQL or SQLite: repo metadata
Summary:
The database access logic is split over several diffs. This initial one contains the overall docblock — read that to understand the rationale for this database.

The current diff introduces the basic database access framework, and the ability to store `repomd.xml` files in the DB (this file is the root index of an RPM repo, pointing at all other data therein).

Future diffs will introduce the ability to record and query "repodata" databases, as well as actual RPM files.

Reviewed By: zeroxoneb

Differential Revision: D12879982

fbshipit-source-id: 7124ab82d68ed20b02ce954a05a1fa5ef3cbf95d