Database access module: RPM storage table
Summary:
This table is very similar to the `Repodata` table, and in fact the same `maybe_store` method is used to write to it.

The one difference is that it has an additional `canonical_checksum` column, which lets us detect the buggy scenario when a given RPM filename (i.e. name-version-release-architecture) occurs in the repo with multiple variants of the content.

The current diff does NOT do anything with canonical checksums, that's the next part of this stack.

Reviewed By: zeroxoneb

Differential Revision: D15457913

fbshipit-source-id: 8f2f1c3d256368fb3e1de88a6de89a94b96156b2