Introduce `temp_repos.py` to stop committing test RPMs to source control
Summary:
This is essentially a copy of the existing `build_repos.py`, which is designed to run at **test-time**. The idea is to avoid committing any RPMs or RPM repos into source control at all.

Follow-ups D16239284, D16239279, and D16239285 migrate RPM tests not to need the in-source-control ":test_repos" target. That allows D16239283 and D16239278 to delete `build_repos.py` and the in-source-tree repo data.

`temp_repos.py` introduces just one change in semantics compared to `build_repos.py`, which is that only the `milk` RPM in one repo contains a `bin/sh` and a post-install script. The goal of changing the semantics in this way is to keep most RPMs as small as possible (and thus keep the snapshot as small as possible), in order to keep all our tests as performant as possible.

This semantics change affects some tests that use test snapshots, so D16239281 fixes up the tests before committing a lighter snapshot to the source tree. The eventual idea is to also generate the snapshots on-the-fly, but that's for later.

The net effect of generating repos on-the-fly, and only having 1 RPM with busybox in it is that our `rpm/tests` directory shrinks from 11MB to 1.4MB.

Reviewed By: zeroxoneb

Differential Revision: D16239282

fbshipit-source-id: 4b1dc8f2691bfcc6f215de22f6c84cff4f248b0f