Fix test_ensure_demo_sendstreams_cover_all_operations on newer kernels
Summary:
Older kernel versions used to emit superfluous `truncate` operations in a number of scenarios, but e.g. 4.16 (and possibly earlier), the only ways I know to get a `truncate` are the times it's truly needed, so let's correctly exercise that functionality:
 - Add a hole at the end of the file ("a sparse file where the inode size extends past the written size" in the words of osandov).
 - Truncate a file in a snapshot, and view an incremental sendstream relative to the parent.

This re-records the sendstream with a newer kernel for those tests that are pinned to our "gold" sendstream.

This update that nothing else changed in the sendstream format since then (as far as our test-cases are concerned).

Reviewed By: zeroxoneb

Differential Revision: D16444882

fbshipit-source-id: 8b89a45eca49c26e84887fa0080f1f550c50f8f1