Support systemd-nspawn version 242 and the new --console=pipe
Summary:
Support systemd-nspawn version 242 and the new --console=pipe
argument (or --pipe for short).  With systemd-nspawn version 242 a new
--console= argument was added that properly handles stderr/stdout file
descriptors when not invoked interactively.  Let's use it.

I don't like that I have to check the systemd-nspawn version,but I don't know
how else to do this and I suspect that as we move into the future we'll have to
do more stuff like this as features are added to nspawn.

Reviewed By: snarkmaster

Differential Revision: D15812001

fbshipit-source-id: 5283e129c8db8704c1793d43f70f78fc172f72e7