Further refactor of image unittest helpers
Summary:
This is a follow-up to sameetandpotatoes's D16605375 to make the helper interface usable for my upcoming `image.cpp_unittest`.

The biggest change in this diff is an update in the comment narrative around the so-called "porcelain deps". Now that we are doing C++, they seem likely to be a permanent feature of this macro library. Read the details inline.

I also fixed some naming nits, and updated some stale "fake_macro_library" definitions.

Reviewed By: zeroxoneb

Differential Revision: D16635281

fbshipit-source-id: efdf8fbbf5741dc9044cd440ac5827ea436c5709