the autodeps codebase
Summary:
There are still a few inverted deps,  python_resolver is inverted.  So we need to break autodeps into a lib target.

I added autodeps-skip to those targets that had bogus deps added because of test contents

And to `add_unowned_headers` and `expand_auto_headers` which tried to remove a bunch of deps and I was scared.

Reviewed By: yfeldblum, luciang

Differential Revision: D16576725

fbshipit-source-id: 5d35ef2c82cba5e44d139d8326905d1edf162ca7