Improve procfs_serde API, fix related bug
Summary:
In debugging RPM issues with anitazha, we found that the repo was getting bind-mounted even in mode/opt.

The cause of the bug was of course that `'0'` is truthy.

I feel appropriately bad for making `procfs_serde` an untyped API — reviewers: do we want (in a future, more disruptive diff) to type-mark all files in the system? A mandatory short suffix on the filename would be fine by me. E.g. `artifacts_may_require_repo=i`.

Reviewed By: zeroxoneb

Differential Revision: D16662988

fbshipit-source-id: 807092ddb6afd0f13f7d299b2b748eb921711dda