fs_image: helpers for rpm version comparison
Summary:
2 ways of generating `RpmMetdata`, which holds RPM versioning info: one for
reading the version info of an installed RPM out of a subvolume, and one for reading
from a local RPM file.

Yum RPM version comparison uses the `rpm` module as RPM shown at
https://github.com/rpm-software-management/yum/blob/master/rpmUtils/miscutils.py.
The `rpmdev-vercmp` tool also uses this module.

I did implement a C -> Python of how RPM compares versions, but the special
handling of ~ and ^ were not implemented in the `rpm` Python module and hence seems
irrelevant if Yum is our primary source of truth.

Reviewed By: snarkmaster

Differential Revision: D16554159

fbshipit-source-id: f4430fd88769545d0ae2410ccabc6cbb839fd70a