install local rpm into image.layer
Summary:
Support for installing local RPM files into an image layer.

We want the specific RPM installed, no matter what's already on the system. This is the sanest way to e.g. replace a base `systemd` with a development `systemd` — remove & re-install is not a good option since that requires breaking dependencies.

There is no all-in-one verb that supersedes `install`, `upgrade`, and `downgrade` to just forcibly install the file specified by the user. `localinstall` looked promising, but it does not handle downgrades (it just skips those packages).

When installing a local RPM, we get the version of that RPM installed in the parent layer (if it was even installed). Then we use `RpmMetadata` to compare the versions of the local RPM and the version in the parent layer to choose whether to keep the Yum verb as "install" or whether to use "downgrade".

Reviewed By: snarkmaster

Differential Revision: D16436679

fbshipit-source-id: 4ef7cd27350f36d362b7d168db1dde0214a3290e