Tests use an ephemeral snapshot of the test repos
Summary:
As a temporary hack, before the RPM snapshotting stack was committed, many of our `yum`-related tests indirectly used a checked-in snapshot in `fs_image/rpm/tests/snapshot`. This weighed 1.3MB (mostly because of `busybox`), and needed to be updated whenever something about the snapshotting toolchain changed.

This change defines the logic to generate the snapshot (which is similar to how the old in-repo snapshot was made), and runs it as part of the build. This guarantees that the snapshot is never stale, making development easier.

Rebuilding the snapshot every time does have a cost, but it is mostly masked by Buck's parallel build. My unscientific experiment suggests that the difference is close to the run-to-run noise on my system, but at most ~7 seconds cold, and ~1 second warm.

- Before this change:

```
$ buck clean && buck test //fs_image/...  # cold tests
Building: finished in 01:22.2 min (100%) 7273/7273 jobs, 7273 updated
  Total time: 01:39.6 min
...
Summary (total time 40.18s):
  PASS: 311

$ buck test //fs_image/...  # warm tests
Summary (total time 31.91s):
  PASS: 311
```

- After this change:

```
$ buck clean && buck test //fs_image/...  # cold tests
Building: finished in 01:28.2 min (100%) 7297/7297 jobs, 7297 updated
  Total time: 01:46.5 min
...
Summary (total time 39.58s):
  PASS: 311

$ buck test //fs_image/...  # warm tests
...
Summary (total time 32.84s):
  PASS: 311
```

Reviewed By: zeroxoneb

Differential Revision: D12879964

fbshipit-source-id: a6ffe705c2cb643f4d3629973e66e50bb9b58b0f