Clean up `image.{layer,feature}` API
Summary:
This introduces the concept of "inlinable" targets, which I plan to use in other parts of the syntax. Basically, we have functions that can (a) take `name` and act as regular targets, or (b) not take `name`, and return something that is functionally equivalent to a target, but does not incur the build overhead.

The changes are as follows:

 - `image.feature` acquires an inline mode, where (roughly) it returns what it would have written to a JSON file if it were made a target.
 - The `features` keyword now takes target paths **or** inlined `image.feature`s.
 - `image.layer` no longer takes **all** `image.feature` kwargs, but only takes `features`. First, this was a bit promiscuous, but more importantly, this made it much harder for people to discover that features are a thing. **The main point of this diff is to be an extended ad for the compositionality of our toolchain.**
 - While I was in there, I cleaned up the `image.layer` code structure to clearly define the split between "built" and "from sendstream" images. This motivated the introduction of the `build_opts` argument.

# Question for reviewers

My least favorite thing about this diff is that it makes the syntax more verbose (and oftentimes, more nested). I think the extra clarity about what's going on is well-worth it. I say this, because I see a few easy opportunities to make the syntax a bit more concise:

Please vote on the below proposals. Other ideas are welcome!

 - **[elide 1-element lists]** Make it so that any place that accepts a list does type inspection, and automatically promotes non-list types to single-element lists. I like how it looks, but I'm not sure if the polymorphic syntax is going to confuse people. Compare:

```
# Before
features = [image.feature(rpms = ["bar"])],
# After
features = image.feature(rpms = "bar"),
```

 - **[shorten option names]** Specifically, `s/build_appliance/appliance/` and `s/yum_from_repo_snapshot/yum/`. Being wrapped inside a `build_opts` object, the context clarifies the intent.

Reviewed By: zeroxoneb

Differential Revision: D16750480

fbshipit-source-id: 080831b402a0b93934a0961c8bbbc4b94fe573a5