Merge `ImageBuildArgs` with `LayerOpts`
Summary:
The two structures have the same purpose, and it was only an accident that `Item.build` implementations used a different subset of flags than `Item.get_phase_builder` implementations.

In an upcoming diff, I wanted both structs in one place, and then it became clear it's cleaner to just merge them. And indeed, this diff has a net negative line count.

Reviewed By: zeroxoneb

Differential Revision: D16818859

fbshipit-source-id: 23ad9fc20ec7b015d4ebb1a3add2a9235778fa17