Split filesystem utilities out of rpm/common.py
Summary:
This is a straight code move.

The end goal is to use `Path` throughout the code (caveat: our on-disk serialization format is JSON, so there's something to do there, potentially). Therefore, it should not be in the `rpm` submodule.

The other utilities also seem unrelated to RPMs, so I pulled them out too.

Reviewed By: zeroxoneb

Differential Revision: D16939494

fbshipit-source-id: 945e12232f3254fca3dd4b39faf977ab90aeea43